+++ src/main/java/net/geforcemods/securitycraft/blockentities/ReinforcedHopperBlockEntity.java
@@ -1,5 +1,7 @@
 package net.geforcemods.securitycraft.blockentities;
 
+import java.util.EnumMap;
+
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IOwnable;
@@ -18,6 +20,7 @@ import net.minecraft.world.level.block.state.BlockState;
 public class ReinforcedHopperBlockEntity extends HopperBlockEntity implements IOwnable, IModuleInventory {
 	private NonNullList<ItemStack> modules = NonNullList.withSize(getMaxNumberOfModules(), ItemStack.EMPTY);
 	private Owner owner = new Owner();
+	private EnumMap<ModuleType, Boolean> moduleStates = new EnumMap<>(ModuleType.class);
 
 	public ReinforcedHopperBlockEntity(BlockPos pos, BlockState state) {
 		super(pos, state);
@@ -34,6 +37,7 @@ public class ReinforcedHopperBlockEntity extends HopperBlockEntity implements IO
 
 		owner.load(tag);
 		modules = readModuleInventory(tag);
+		moduleStates = readModuleStates(tag);
 	}
 
 	@Override
@@ -44,6 +48,7 @@ public class ReinforcedHopperBlockEntity extends HopperBlockEntity implements IO
 			owner.save(tag, false);
 
 		writeModuleInventory(tag);
+		writeModuleStates(tag);
 		return tag;
 	}
 
@@ -98,5 +103,20 @@ public class ReinforcedHopperBlockEntity extends HopperBlockEntity implements IO
 	@Override
 	public NonNullList<ItemStack> getInventory() {
 		return modules;
+	}
+
+	@Override
+	public boolean isModuleEnabled(ModuleType module) {
+		return hasModule(module) && moduleStates.get(module);
+	}
+
+	@Override
+	public void toggleModuleState(ModuleType module, boolean shouldBeEnabled) {
+		moduleStates.put(module, shouldBeEnabled);
+
+		if (shouldBeEnabled)
+			onModuleInserted(getModule(module), module, true);
+		else
+			onModuleRemoved(getModule(module), module, true);
 	}
 }