+++ src/main/java/net/geforcemods/securitycraft/network/client/RefreshDisguisableModel.java
@@ -15,19 +15,22 @@ public class RefreshDisguisableModel {
 	private BlockPos pos;
 	private boolean insert;
 	private ItemStack stack;
+	private boolean toggled;
 
 	public RefreshDisguisableModel() {}
 
-	public RefreshDisguisableModel(BlockPos pos, boolean insert, ItemStack stack) {
+	public RefreshDisguisableModel(BlockPos pos, boolean insert, ItemStack stack, boolean toggled) {
 		this.pos = pos;
 		this.insert = insert;
 		this.stack = stack;
+		this.toggled = toggled;
 	}
 
 	public static void encode(RefreshDisguisableModel message, FriendlyByteBuf buf) {
 		buf.writeBlockPos(message.pos);
 		buf.writeBoolean(message.insert);
 		buf.writeItem(message.stack);
+		buf.writeBoolean(message.toggled);
 	}
 
 	public static RefreshDisguisableModel decode(FriendlyByteBuf buf) {
@@ -36,6 +39,7 @@ public class RefreshDisguisableModel {
 		message.pos = buf.readBlockPos();
 		message.insert = buf.readBoolean();
 		message.stack = buf.readItem();
+		message.toggled = buf.readBoolean();
 		return message;
 	}
 
@@ -45,9 +49,9 @@ public class RefreshDisguisableModel {
 
 			if (be != null) {
 				if (message.insert)
-					be.insertModule(message.stack);
+					be.insertModule(message.stack, message.toggled);
 				else
-					be.removeModule(ModuleType.DISGUISE);
+					be.removeModule(ModuleType.DISGUISE, message.toggled);
 
 				ClientHandler.refreshModelData(be.getBlockEntity());
 			}